+++ CarSales/LeadsUserRights.aspx.cs
@@ -466,14 +466,22 @@ public partial class LeadsUserRights : System.Web.UI.Page
         chkleadsadmin.Items[0].Selected = false;
 
     }
-    public void btndelete_Click(object sender, EventArgs e)
+    public void OnConfirm(object sender, EventArgs e)
     {
+        string confirmValue = Request.Form["confirm_value"];
+        if (confirmValue == "Yes")
+        {
 
-        //Deactivating Empid from carsales.
-        string EMpid = Session["EMpid"].ToString();
-        DataSet dsSalesUpdateList = objHotLeadBL.DeleteEmployee(EMpid);
-        System.Web.UI.ScriptManager.RegisterClientScriptBlock(Page, typeof(Page), "Script", "alert('" + EMpid + " has been deleted succesfully.');", true);
-        MpUpdaterights.Hide();
-        GetLeadsUserRights(Convert.ToInt32(ddlcenters.SelectedValue));
+            //Deactivating Empid from carsales.
+            string EMpid = Session["EMpid"].ToString();
+            DataSet dsSalesUpdateList = objHotLeadBL.DeleteEmployee(EMpid);
+            System.Web.UI.ScriptManager.RegisterClientScriptBlock(Page, typeof(Page), "Script", "alert('" + EMpid + " has been deleted succesfully.');", true);
+            MpUpdaterights.Hide();
+            GetLeadsUserRights(Convert.ToInt32(ddlcenters.SelectedValue));
+        }
+        else
+        {
+
+        }
     }
 }
